+++ trunk/ssf/main/src/org/kablink/teaming/portlet/administration/ManageGroupPrincipalsController.java
@@ -63,6 +63,7 @@ import org.kablink.teaming.util.LongIdUtil;
 import org.kablink.teaming.web.WebKeys;
 import org.kablink.teaming.web.portlet.SAbstractController;
 import org.kablink.teaming.web.util.PortletRequestUtils;
+import org.kablink.teaming.web.util.WebHelper;
 import org.kablink.util.Validator;
 import org.kablink.util.search.Constants;
 import org.springframework.web.portlet.ModelAndView;
@@ -73,7 +74,7 @@ public abstract class ManageGroupPrincipalsController extends  SAbstractControll
 		Map formData = request.getParameterMap();
 		Long binderId = PortletRequestUtils.getRequiredLongParameter(request, WebKeys.URL_BINDER_ID);
 		response.setRenderParameter(WebKeys.URL_BINDER_ID, binderId.toString());
-		if (formData.containsKey("addBtn")) {
+		if (formData.containsKey("addBtn") && WebHelper.isMethodPost(request)) {
 			
 			String name = PortletRequestUtils.getStringParameter(request, "name", "").trim();
 			
@@ -97,7 +98,7 @@ public abstract class ManageGroupPrincipalsController extends  SAbstractControll
 				response.setRenderParameter(WebKeys.EXCEPTION, iae.getLocalizedMessage());
 			}
 			
-		} else if (formData.containsKey("applyBtn") || formData.containsKey("okBtn")) {
+		} else if ((formData.containsKey("applyBtn") || formData.containsKey("okBtn")) && WebHelper.isMethodPost(request)) {
 			Long groupId = PortletRequestUtils.getRequiredLongParameter(request, WebKeys.URL_ENTRY_ID);
 			String title = PortletRequestUtils.getStringParameter(request, "title", "");
 			String description = PortletRequestUtils.getStringParameter(request, "description", "");
@@ -111,7 +112,7 @@ public abstract class ManageGroupPrincipalsController extends  SAbstractControll
 			getProfileModule().modifyEntry( groupId, new MapInputData(updates));
 			response.setRenderParameter(WebKeys.URL_ENTRY_ID, groupId.toString());
 
-		} else if (formData.containsKey("deleteBtn")) {
+		} else if (formData.containsKey("deleteBtn") && WebHelper.isMethodPost(request)) {
 			Long groupId = PortletRequestUtils.getRequiredLongParameter(request, WebKeys.URL_ENTRY_ID);
 			getProfileModule().deleteEntry(groupId, null);
